add IDamagable interface

+++ Assets/Scripts/Interface/IDamagable.cs
@@ -2,5 +2,5 @@
 
 public interface IDamagable
 {
-    public void TakeDamage(float amount);
+    public void TakeDamage(int amount);
 }
